Move dependency checker code to new class
